cabana: move RoutesDialog out of streams/
